Event tab, start of componentizing the whole app, fix trace click bug
