+++ web/app/src/app/app.module.ts
@@ -58,6 +58,9 @@ import { MonacoEditorModule } from 'ngx-monaco-editor';
 import { SettingsComponent } from './settings/settings.component';
 import { EventsComponent } from './events/events.component';
 import { DateAgoPipe } from './dateago.pipe';
+import { EventsListComponent } from './events-list/events-list.component';
+import { StatChartsComponent } from './stat-charts/stat-charts.component';
+import { TraceListComponent } from './trace-list/trace-list.component';
 
 /**
  * Import specific languages to avoid importing everything
@@ -87,7 +90,10 @@ export function getHighlightLanguages() {
     NotInvitedComponent,
     SettingsComponent,
     EventsComponent,
-    DateAgoPipe
+    DateAgoPipe,
+    EventsListComponent,
+    StatChartsComponent,
+    TraceListComponent
   ],
   imports: [
     BrowserModule,
