+++ web/app/src/app/events-list/events-list.component.html
@@ -1,0 +1,21 @@
+<mat-card
+  class="mat-card-flat"
+  *ngFor="let event of events"
+  style="margin-bottom: 1em; background: #f5f5f5;"
+>
+  <mat-card-title style="align-items: center; font-size: 1.25em;">
+    <div fxLayout.lt-md="column" fxLayout="row">
+      <div>
+        <a routerLink="/service/{{ event.service.name }}"
+          >{{ event.service.name }}:</a
+        >
+        &nbsp;
+      </div>
+      <div>
+        {{ eventTypeToString(event) }} &nbsp;<span style="color: #555">{{
+          event.timestamp * 1000 | dateAgo
+        }}</span>
+      </div>
+    </div>
+  </mat-card-title>
+</mat-card>
