+++ web/app/src/app/events/events.component.ts
@@ -2,20 +2,6 @@ import { Component, OnInit } from "@angular/core";
 import { ServiceService } from "../service.service";
 import * as types from "../types";
 
-const eventTypes = {
-  1: "DepoymentCreated",
-  2: "DepoymentDeleted",
-  3: "DepoymentUpdated",
-  4: "SourceUpdated"
-};
-
-const eventTypesNice = {
-  1: "deployment created",
-  2: "deployment deleted",
-  3: "deployment updated",
-  4: "source updated"
-};
-
 @Component({
   selector: "app-events",
   templateUrl: "./events.component.html",
@@ -32,7 +18,4 @@ export class EventsComponent implements OnInit {
     });
   }
 
-  eventTypeToString(e: types.Event): string {
-    return eventTypesNice[e.type];
-  }
 }
