+++ web/app/src/app/stat-charts/stat-charts.component.html
@@ -1,0 +1,66 @@
+<div
+  fxLayout.lt-md="column wrap"
+  fxLayout="row wrap"
+  fxLayoutAlign="space-between"
+>
+  <div fxFlex="48%" style="margin-bottom: 2em;">
+    <canvas
+      *ngIf="requestRates.data.length > 0"
+      baseChart
+      [colors]="requestRates.chartColors"
+      [datasets]="requestRates.data"
+      [options]="requestRates.options"
+      [chartType]="requestRates.lineChartType"
+    ></canvas>
+  </div>
+  <div fxFlex="48%" style="margin-bottom: 2em; ">
+    <canvas
+      *ngIf="errorRates.data.length > 0"
+      baseChart
+      [colors]="errorRates.chartColors"
+      [datasets]="errorRates.data"
+      [options]="errorRates.options"
+      [chartType]="errorRates.lineChartType"
+    ></canvas>
+  </div>
+  <div fxFlex="48%" style="margin-bottom: 2em; ">
+    <canvas
+      *ngIf="memoryRates.data.length > 0"
+      baseChart
+      [colors]="memoryRates.chartColors"
+      [datasets]="memoryRates.data"
+      [options]="memoryRates.options"
+      [chartType]="memoryRates.lineChartType"
+    ></canvas>
+  </div>
+  <div fxFlex="48%" style="margin-bottom: 2em; ">
+    <canvas
+      *ngIf="concurrencyRates.data.length > 0"
+      baseChart
+      [colors]="concurrencyRates.chartColors"
+      [datasets]="concurrencyRates.data"
+      [options]="concurrencyRates.options"
+      [chartType]="concurrencyRates.lineChartType"
+    ></canvas>
+  </div>
+  <div fxFlex="48%" style="margin-bottom: 2em; ">
+    <canvas
+      *ngIf="gcRates.data.length > 0"
+      baseChart
+      [colors]="gcRates.chartColors"
+      [datasets]="gcRates.data"
+      [options]="gcRates.options"
+      [chartType]="gcRates.lineChartType"
+    ></canvas>
+  </div>
+  <div fxFlex="48%" style="margin-bottom: 2em;">
+    <canvas
+      *ngIf="uptime.data.length > 0"
+      baseChart
+      [colors]="uptime.chartColors"
+      [datasets]="uptime.data"
+      [options]="uptime.options"
+      [chartType]="uptime.lineChartType"
+    ></canvas>
+  </div>
+</div>
