+++ web/app/src/app/trace-list/trace-list.component.html
@@ -1,0 +1,50 @@
+<div fxLayout="row">
+  <div style="width: 25%">
+    <b style="font-weight: 700;">ID</b>
+  </div>
+  <div style="width: 25%">
+    <b style="font-weight: 700;">Name</b>
+  </div>
+  <div style="width: 25%">
+    <b style="font-weight: 700;">Time</b>
+  </div>
+  <div style="width: 25%">
+    <b style="font-weight: 700;">Duration</b>
+  </div>
+</div>
+<br />
+<ng-container *ngFor="let traceData of traceDatasPart">
+  <div fxLayout="row">
+    <div style="width: 25%">
+      <a href="#" (click)="show(traceData)"
+        >{{ prettyId(traceData.traceId) }}
+      </a>
+    </div>
+    <div style="width: 25%">
+      {{ getEndpointName(traceData.dataTable) }}
+    </div>
+    <div style="width: 25%">
+      {{ traceData.dataTable[1][2] | date: "short" }}
+    </div>
+    <div style="width: 25%">
+      {{ traceDuration(traceData.dataTable) }}
+    </div>
+  </div>
+  <div
+    *ngIf="traceData.show"
+    style="width: 100%;"
+    [ngStyle]="{ 'height.px': traceData.divHeight + 25 }"
+  >
+    <br />Trace ID: {{ traceData.traceId }}<br />
+    <google-chart [data]="traceData"></google-chart>
+  </div>
+</ng-container>
+<mat-paginator
+  #paginator
+  [length]="length"
+  [pageIndex]="currentPage"
+  [pageSize]="pageSize"
+  [pageSizeOptions]="[5, 10, 25, 100]"
+  (page)="handlePage($event)"
+>
+</mat-paginator>
